+++ zopkio/__init__.py
@@ -17,6 +17,6 @@
 # specific language governing permissions and limitations
 # under the License.
 
-"""DTF test framework"""
+"""Zopkio test framework"""
 
 __version__ = '0.1.0'
