+++ zopkio/adhoc_deployer.py
@@ -23,9 +23,9 @@ import tarfile
 import time
 import zipfile
 
-import dtf.constants as constants
-from dtf.deployer import Deployer, Process
-from dtf.remote_host_helper import better_exec_command, DeploymentError, get_sftp_client, get_ssh_client, open_remote_file
+import zopkio.constants as constants
+from zopkio.deployer import Deployer, Process
+from zopkio.remote_host_helper import better_exec_command, DeploymentError, get_sftp_client, get_ssh_client, open_remote_file
 
 logger = logging.getLogger(__name__)
