+++ zopkio/web_resources/footer.html
@@ -1,6 +1,6 @@
     <div class="container">
       <footer>
-        <p>dtf &copy; LinkedIn 2014</p>
+        <p>zopkio &copy; LinkedIn 2014</p>
       </footer>
     </div>
   </body>
