+++ zopkio/web_resources/logs_page.html
@@ -2,7 +2,7 @@
   <div class="topbar navbar navbar-default navbar-fixed-top">
     <div class="navbar-inner">
       <div class="container">
-        <a class="brand navbar-brand" href="{{ report_info.project_url }}">dtf</a>
+        <a class="brand navbar-brand" href="{{ report_info.project_url }}">zopkio</a>
         <ul class="nav navbar-nav">
           <li><a href="{{ report_info.home_page }}"><span>Summary</span></a></li>
           <li class="dropdown active">
