+++ zopkio/web_resources/topbar.html
@@ -1,7 +1,7 @@
 <div class="topbar navbar navbar-default navbar-fixed-top">
     <div class="navbar-inner">
       <div class="container">
-        <a class="brand navbar-brand" href="{{ report_info.project_url }}">dtf</a>
+        <a class="brand navbar-brand" href="{{ report_info.project_url }}">zopkio</a>
         <ul class="nav navbar-nav">
           {%- if active == "summary" %}
             <li class="active">
